debugging.

+++ bsp/gpio.c
@@ -194,7 +194,7 @@ void gpio_set_event_detect(gpio_pin_t pin, event_type_t event_type)
 			break;
 		case FALLING_EDGE_DETECT:
 			printf("%s:%d - \n", __FILE__, __LINE__);
-			printf("GPIO_BAR=%08X   &GPIO_BAR[idx]=%08X\n", GPIO_BAR, &GPIO_BAR[word+(GPFEN0-GPREN0)]);
+			printf("GPIO_BAR=%08X   &GPIO_BAR[idx]=%08X\n", (unsigned)GPIO_BAR, (unsigned)&GPIO_BAR[word+(GPFEN0-GPREN0)]);
 			printf("GPIO_BAR[word+(GPFEN0-GPREN0)]=%08X\n", GPIO_BAR[word+(GPFEN0-GPREN0)]);
 			printf("(1<<pinInWord)=%08X\n", (1<<pinInWord));
 			GPIO_BAR[word+(GPFEN0-GPREN0)] |= (1<<pinInWord);
